Flat rides: widen bounding-box lengths to uint16 (tall-ride sort)
Companion to the uint16 invalidation fix. FlatRideRotationDescriptor.BbLength{X,Y,Z}
were uint8_t (max 255), so the painter's-algorithm sort box for a ride could never be
taller than 255 units - far short of a ~648px high-res Ferris Wheel, whose upper
portion therefore still lost the depth-sort against the tiles behind it and clipped
even with bbLengthZ maxed. (bbLengthZ=200 wasn't capped, just too short; the ceiling
made a tall-enough box impossible.)

Widened to uint16. Like the Invalidation* fields, BbLength* come from the ride
object's JSON, NOT the park save (ParkFile), so this changes no save format. Defaults
stay 24/24/48; short rides are unaffected. Authored via AttractionEditorTool (Bounding
box Z now accepts several hundred).

+++ src/openrct2/object/RideObject.cpp
@@ -1263,10 +1263,11 @@ namespace OpenRCT2
         data->Descriptor.FramesPerDir            = Json::GetNumber<uint16_t>(j["framesPerDir"]);
         data->Descriptor.StructureZOffset        = Json::GetNumber<int8_t>(j["structureZOffset"], 7);
         // Paint bounding-box (sort volume); BbLengthZ must reach a tall ride's top or it clips.
-        // Defaults match the legacy hardcoded 24/24/48 so untouched rides render identically.
-        data->Descriptor.BbLengthX               = Json::GetNumber<uint8_t>(j["bbLengthX"], 24);
-        data->Descriptor.BbLengthY               = Json::GetNumber<uint8_t>(j["bbLengthY"], 24);
-        data->Descriptor.BbLengthZ               = Json::GetNumber<uint8_t>(j["bbLengthZ"], 48);
+        // uint16 so a large high-res ride can exceed 255. Defaults match the legacy hardcoded
+        // 24/24/48 so untouched rides render identically.
+        data->Descriptor.BbLengthX               = Json::GetNumber<uint16_t>(j["bbLengthX"], 24);
+        data->Descriptor.BbLengthY               = Json::GetNumber<uint16_t>(j["bbLengthY"], 24);
+        data->Descriptor.BbLengthZ               = Json::GetNumber<uint16_t>(j["bbLengthZ"], 48);
         // uint16 (not uint8): a large high-res sprite needs a redraw region past 255px - see
         // FlatRideRotationDescriptor and Vehicle::UpdateFlatRideGeneric's direct invalidation.
         data->Descriptor.InvalidationHalfWidth   = Json::GetNumber<uint16_t>(j["invalidationHalfWidth"]);
